- Reduced precision for gamma, thm, thp, and other constants, however it does not seem to have much impact on performance, replaced pow function by multiplication in second order reconstructions

+++ src/bc.h
@@ -8,7 +8,7 @@
 void bc(int ni, int nj, double Q[ni-1][nj-1][4],			\
 	double rho[ni+1][nj+1], double u[ni+1][nj+1], double v[ni+1][nj+1], double p[ni+1][nj+1], \
 	double neta_x[ni-1][nj], double neta_y[ni-1][nj],		\
-	struct state freestream, double gamma=1.4){
+	struct state freestream, float gamma=1.4){
   
   int nim = ni - 1;
   int njm = nj - 1;
