+++ src/flux.h
@@ -1,6 +1,6 @@
 #ifndef _INCL_FLUX
 #define _INCL_FLUX
-#include <cmath>
+#include "mkl.h"
 #include <stdlib.h>
 #include<iostream>
 #include <assert.h>
@@ -13,14 +13,14 @@
 __declspec(vector)  void roeflux(double nx, double ny,		\
 				 double rlft, double ulft, double vlft, double plft, \
 				 double rrht, double urht, double vrht, double prht, \
-				 double gamma, double f[4]){
-  double gm1 = gamma - 1.0;
-  double ogm1 = 1/gm1;
+				 float gamma, double f[4]){
+  float gm1 = gamma - 1.0;
+  float ogm1 = 1/gm1;
 
   double rlfti = 1/rlft;
   double rulft = rlft*ulft;
   double rvlft = rlft*vlft;
-  double uvl = 0.5*(ulft*ulft + vlft*vlft);
+  double uvl = 0.5f*(ulft*ulft + vlft*vlft);
   double elft = plft*ogm1 + rlft*uvl;
   double hlft = (elft + plft)*rlfti;
 
@@ -28,7 +28,7 @@ __declspec(vector)  void roeflux(double nx, double ny,		\
   double rrhti = 1/rrht;
   double rurht = rrht*urht;
   double rvrht = rrht*vrht;
-  double uvr = 0.5*(urht*urht + vrht*vrht);
+  double uvr = 0.5f*(urht*urht + vrht*vrht);
   double erht = prht*ogm1 + rrht*uvr;
   double hrht = (erht + prht)*rrhti;
   
@@ -38,7 +38,7 @@ __declspec(vector)  void roeflux(double nx, double ny,		\
   double uav = (rat*urht + ulft)*rati;
   double vav = (rat*vrht + vlft)*rati;
   double hav = (rat*hrht + hlft)*rati;
-  double uv = 0.5*(uav*uav + vav*vav);
+  double uv = 0.5f*(uav*uav + vav*vav);
   double cav = sqrt(gm1*(hav - uv));
 
   double aq1 = rrht - rlft;
@@ -50,7 +50,7 @@ __declspec(vector)  void roeflux(double nx, double ny,		\
   double r1 = nx/dr;
   double r2 = ny/dr;
 
-  //assert(abs(r2) < 1e-10);
+
   double uu = r1*uav + r2*vav;
   double c2 = cav*cav;
   double c2i = 1/c2;
@@ -60,14 +60,9 @@ __declspec(vector)  void roeflux(double nx, double ny,		\
 
   double uulft = r1*ulft + r2*vlft;
   double uurht = r1*urht + r2*vrht;
-  /* std::cout<<typeid(std::abs(rlft-rrht)).name()<<"\n"; */
-  /* if(abs(rlft-rrht)>1e-10){ */
-  /*   std::cout<<r1<<"uulft "<<vlft<<"\n"; */
-  /*   std::cout<<r1<<"uurht "<<vlft<<"\n"; */
-  /* } */
   double rcav = rav*cav;
   double aquu = uurht - uulft;
-  double c2ih = 0.5*c2i;
+  double c2ih = 0.5f*c2i;
   double ruuav = auu*rav;
   
   double b1, b2, b3, b4, b5, b6, b7;
@@ -84,15 +79,13 @@ __declspec(vector)  void roeflux(double nx, double ny,		\
   aq3 = vav*b4 + r2*b5 + b7;
   aq4 = hav*b4 + uu*b5 + uav*b6 + vav*b7 - c2*b1*ogm1;
 
-  double aj = 0.5*dr;
+  double aj = 0.5f*dr;
   double plar = plft + prht;
   double eplft = elft + plft;
   double eprht = erht + prht;
   f[0] = aj*(rlft*uulft + rrht*uurht - aq1);
   f[1] = aj*(rulft*uulft + rurht*uurht + r1*plar - aq2);
   f[2] = aj*(rvlft*uulft + rvrht*uurht + r2*plar - aq3);
-
-  //  std::cout<<"roe"<<f[2]<<" "<<b7<<"\n";
   f[3] = aj*(eplft*uulft + eprht*uurht - aq4);
 
 }
